+++ admin/webapp/pages/Site/RemovePage.class.php
@@ -6,9 +6,9 @@ class RemovePage extends CMSUpdatePageBase{
 
 	function doPost(){
 
-		if(isset($_POST["confirm"]) && $_POST["confirm"]){
+		if(soy2_check_token() && isset($_POST["confirm"]) && $_POST["confirm"]){
 
-			$action = SOY2ActionFactory::createInstance("Site.RemoveAction",array(
+			$action = SOY2ActionFactory::createInstance("Site.RemoveAction", array(
 	    		"id" => $this->id,
 	    		"deleteDir"      => isset($_POST["deleteDir"])      && $_POST["deleteDir"],
 	    		"deleteDatabase" => isset($_POST["deleteDatabase"]) && $_POST["deleteDatabase"],
@@ -27,56 +27,57 @@ class RemovePage extends CMSUpdatePageBase{
 	}
 
     function RemovePage($args) {
-    	if(!UserInfoUtil::isDefaultUser() || count($args)<1){
+    	if(!UserInfoUtil::isDefaultUser() || count($args) < 1){
     		//デフォルトユーザのみ削除可能
     		$this->jump("Site");
     		exit;
     	}
 
-    	$this->id = $args[0];
+    	$this->id = (isset($args[0])) ? $args[0] : null;
 
     	WebPage::WebPage();
 
 		$SiteLogic = SOY2Logic::createInstance("logic.admin.Site.SiteLogic");
 		$site = $SiteLogic->getById($this->id);
 
-		if(!$site||$site->getSiteType()==Site::TYPE_SOY_SHOP){
+		if(!$site || $site->getSiteType() == Site::TYPE_SOY_SHOP){
 			$this->jump("Site");
 		}
 
 		$this->addForm("delete_site_form");
 
-		$this->createAdd("site_name_title","HTMLLabel",array(
+		$this->addLabel("site_name_title", array(
 			"text" => $site->getSiteName()
 		));
 
-		$this->createAdd("site_name","HTMLLabel",array(
+		$this->addLabel("site_name", array(
 			"text" => $site->getSiteName()
 		));
 
-		$this->createAdd("site_id","HTMLLabel",array(
+		$this->addLabel("site_id", array(
 			"text" => $site->getSiteId()
 		));
 
-		$this->createAdd("site_url","HTMLLink",array(
+		$this->addLink("site_url", array(
 			"href" => $site->getUrl(),
 			"text" => $site->getUrl(),
 		));
 
-		$this->createAdd("checkbox","HTMLCheckBox",array(
+		$this->addCheckBox("checkbox", array(
 			"label" => CMSMessageManager::get("ADMIN_CONFIRM_DELETE_SITE"),
 			"name"  => "confirm",
 			"value" => 1,
 			"style" => "margin-right:1ex;"
 		));
 
-		$this->createAdd("checkbox_delete_dir","HTMLCheckBox",array(
+		$this->addCheckBox("checkbox_delete_dir", array(
 			"label" => CMSMessageManager::get("ADMIN_DELETE_SITE_DIR"),
 			"name"  => "deleteDir",
 			"value" => 1,
 			"style" => "margin-right:1ex;"
 		));
-		$this->createAdd("checkbox_delete_db","HTMLCheckBox",array(
+		
+		$this->addCheckBox("checkbox_delete_db", array(
 			"label" => CMSMessageManager::get("ADMIN_DELETE_SITE_DB"),
 			"name"  => "deleteDatabase",
 			"value" => 1,
@@ -90,26 +91,25 @@ class RemovePage extends CMSUpdatePageBase{
 			$this->addMessage("ADMIN_DETACH_ROOT_SETTING_BEFORE_DELETE_SITE");
 		}
 
-		$this->createAdd("button","HTMLInput",array(
+		$this->addInput("button", array(
 			"disabled" => $site->getIsDomainRoot(),
 			"value" => CMSMessageManager::get("SOYCMS_DELETE")
 		));
 
 		$this->outputMessage();
 
-    	HTMLHead::addLink("site.edit.css",array(
+    	HTMLHead::addLink("site.edit.css", array(
 			"type" => "text/css",
 			"rel" => "stylesheet",
-			"href" => SOY2PageController::createRelativeLink("./css/site/edit.css")."?".SOYCMS_BUILD_TIME
+			"href" => SOY2PageController::createRelativeLink("./css/site/edit.css") . "?" . SOYCMS_BUILD_TIME
 		));
-
     }
 
     function outputMessage(){
     	$messages = CMSMessageManager::getMessages();
-    	$this->createAdd("error","HTMLLabel",array(
-    		"text" => implode("\n",$messages),
-    		"visible" => !empty($messages)
+    	$this->addLabel("error", array(
+    		"text" => implode("\n", $messages),
+    		"visible" => (!empty($messages))
     	));
     }
 }
